feat(activity_5): add 'IntensiveSearch' and incorporate it into 'TabuSearch'
There is a know error: the list of candidates to
add is being created based on the elements which
fit in the current solution. That is wrong: one
must consider that by removing element(s) more
space will be available so more elements will
fit (i.e. there will possibly bne more candidates).

Currently, that is a logical error in the software.